+++ cmd/lino/dep_dataconnector.go
@@ -31,6 +31,8 @@ func dataPingerFactory() map[string]domain.DataPingerFactory {
 		"postgres":   infra.NewSQLDataPingerFactory(),
 		"godror":     infra.NewSQLDataPingerFactory(),
 		"godror-raw": infra.NewSQLDataPingerFactory(),
+		"mysql":      infra.NewSQLDataPingerFactory(),
 		"db2":        infra.NewSQLDataPingerFactory(),
+		"http":       infra.NewHTTPDataPingerFactory(),
 	}
 }