+++ pkg/table/model.go
@@ -17,10 +17,18 @@
 
 package table
 
+// Column holds the name of a column.
+type Column struct {
+	Name   string
+	Export string
+	Import string
+}
+
 // Table holds a name (table name) and a list of keys (table columns).
 type Table struct {
-	Name string
-	Keys []string
+	Name    string
+	Keys    []string
+	Columns []Column
 }
 
 // Error is the error type returned by the domain